Product info extracted to seperate component
