+++ frontend/src/components/ProductInfo/productInfo.css
@@ -1,0 +1,30 @@
+.product-info-container {
+    display: flex;
+    flex-direction: column;
+    text-align: left;
+    justify-content: space-around;
+}
+
+.place-bid-container {
+    display: flex;
+    flex-wrap: wrap;
+}
+
+.place-bid-form {
+    text-align: center;
+    width: 176px;
+    box-shadow: none;
+    margin-right: 20px;
+}
+
+.place-bid-label {
+    font-size: 15px;
+    color: #9B9B9B;
+    letter-spacing: 0.52px;
+    margin-top: 10px;
+}
+
+.wishlist-button {
+    width: 140px;
+    height: 36px;
+}
